requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
certifi==2023.11.17
click==8.1.7
ecdsa==0.18.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
loguru==0.7.2
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
typing_extensions==4.9.0
uvicorn==0.27.0
uvloop==0.19.0
